refactor:  정적 이벤트 리스너 추가로직을 메서드로 만들고 호출하기
